+++ packages/google_maps_flutter/google_maps_flutter_android/pigeons/messages.dart
@@ -38,7 +38,7 @@ class PlatformCircle {
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
-  final Object json;
+  final Map<String?, Object?> json;
 }
 
 /// Pigeon equivalent of the ClusterManager class.
@@ -57,7 +57,7 @@ class PlatformMarker {
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
-  final Object json;
+  final Map<String?, Object?> json;
 }
 
 /// Pigeon equivalent of the Polygon class.
@@ -69,7 +69,7 @@ class PlatformPolygon {
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
-  final Object json;
+  final Map<String?, Object?> json;
 }
 
 /// Pigeon equivalent of the Polyline class.
@@ -81,7 +81,7 @@ class PlatformPolyline {
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
-  final Object json;
+  final Map<String?, Object?> json;
 }
 
 /// Pigeon equivalent of the TileOverlay class.
@@ -93,7 +93,7 @@ class PlatformTileOverlay {
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
-  final Object json;
+  final Map<String?, Object?> json;
 }
 
 /// Pigeon equivalent of LatLng.
@@ -139,7 +139,7 @@ class PlatformMapConfiguration {
   /// to the internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
-  final Object json;
+  final Map<String?, Object?> json;
 }
 
 /// Pigeon representation of an x,y coordinate.